+++ src/test/java/org/plumelib/StringsPlumeTest.java
@@ -255,6 +255,23 @@ public final class StringsPlumeTest {
   }
 
   @Test
+  public void test_conjunction() {
+
+    // public static String conjunction(String conjunction, List<?> elements)
+
+    assertEquals("a", StringsPlume.conjunction("and", Arrays.asList("a")));
+    assertEquals("a and b", StringsPlume.conjunction("and", Arrays.asList("a", "b")));
+    assertEquals("a, b, and c", StringsPlume.conjunction("and", Arrays.asList("a", "b", "c")));
+    assertEquals(
+        "a, b, c, and d", StringsPlume.conjunction("and", Arrays.asList("a", "b", "c", "d")));
+    assertEquals("a", StringsPlume.conjunction("or", Arrays.asList("a")));
+    assertEquals("a or b", StringsPlume.conjunction("or", Arrays.asList("a", "b")));
+    assertEquals("a, b, or c", StringsPlume.conjunction("or", Arrays.asList("a", "b", "c")));
+    assertEquals(
+        "a, b, c, or d", StringsPlume.conjunction("or", Arrays.asList("a", "b", "c", "d")));
+  }
+
+  @Test
   public void test_rpad() {
 
     // public static String rpad(String s, int length)